Change DPS/PS functions to CGFloat and NSInteger.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@36063 72102866-910b-0410-8b05-ffd578937521

+++ Headers/gsc/GSStreamGState.h
@@ -33,8 +33,8 @@
 {
 @public
   int clinecap, clinejoin;
-  float clinewidth, cmiterlimit;
-  float cstrokeadjust;
+  CGFloat clinewidth, cmiterlimit;
+  CGFloat cstrokeadjust;
 }
 
 @end
